# lib/api-zod/src/generated/types/aiGenerateDraftBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiGenerateDraftBody {
  type: string;
  platform: string;
  topic: string;
  tone?: string;
  additionalContext?: string;
}