# lib/api-zod/src/generated/types/contentPiece.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContentPiece {
  id: number;
  type: string;
  platform: string;
  title: string;
  body: string;
  status: string;
  createdAt: Date;
  updatedAt: Date;
}