# lib/api-zod/src/generated/types/researchNote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResearchNote {
  id: number;
  category: string;
  topic: string;
  content: string;
  createdAt: Date;
  updatedAt: Date;
}