# lib/api-zod/src/generated/types/scheduledPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduledPost {
  id: number;
  platform: string;
  content: string;
  scheduledAt: Date;
  status: string;
  createdAt: Date;
  updatedAt: Date;
}